[refactor] #98 : mp페이지 기능별로 묶기
