+++ src/features/mapView/ui/Detail/TransferDetail.tsx
