+++ src/features/mapView/ui/Detail/UserCard.tsx
