+++ src/features/mapView/ui/Detail/index.ts
@@ -1,0 +1,2 @@
+export * from "./TransferDetail";
+export * from "./UserCard";
