+++ src/features/mapView/ui/Map/DetailKakaoMapView.tsx
